Port pianoteq installation script to install correctly for aarch64 and add zynthbox-dependency-python3-lilv

+++ scripts/install_pianoteq_binary.sh
@@ -11,25 +11,25 @@ pack_fpath=$1
 cd $ZYNTHIAN_SW_DIR
 
 # Save current binary, deleting the older one
-if [ -f pianoteq6/pianoteq ]; then 
+if [ -f pianoteq6/pianoteq ]; then
 	rm -rf pianoteq6.old
 	mv pianoteq6 pianoteq6.old
 else
 	rm -rf pianoteq6
 fi
 
+if [ ${MACHINE_HW_NAME} = "armv7l" ]; then
+	armdir="arm-32bit"
+elif [ ${MACHINE_HW_NAME} = "aarch64" ]; then
+	armdir="arm-64bit"
+fi
+
 # Uncompress new binary package and delete unused files
-7z x "$pack_fpath" \*/arm-32bit \*/arm  \*/Documentation \*/README_LINUX.txt \*/Licence.rtf
+7z x "$pack_fpath" \*/$armdir \*/Documentation \*/README_LINUX.txt \*/Licence.rtf
 mv Pianoteq* pianoteq6
 
 # Create symlink to binary
 cd pianoteq6
-if [ -d arm ]; then
-	armdir="arm"
-else
-	armdir="arm-32bit"
-fi
-
 ln -s ./$armdir/Pianoteq* .
 rm -f *.lv2
 rm -f *.so
@@ -38,12 +38,22 @@ mv Pianoteq* pianoteq
 # Delete old LV2 plugin
 rm -rf $ZYNTHIAN_PLUGINS_DIR/lv2/Pianoteq*.lv2
 
+exit
+
+#----------------------------------------------------------
+# Don't install Pianoteq LV2 to avoid confusion
+#----------------------------------------------------------
+
 # Create symlink to LV2 plugin directory.
 ln -s $ZYNTHIAN_SW_DIR/pianoteq6/$armdir/*.lv2 $ZYNTHIAN_PLUGINS_DIR/lv2
 
 # Generate LV2 presets (Pianoteq >= v7.2.1)
-version=$(./pianoteq --version | cut -d' ' -f4)
-if [[ $version > "7.2.0" ]]; then
-	rm -rf $ZYNTHIAN_MY_DATA_DIR/presets/lv2/"Pianoteq 7 "*"-factory-presets"*.lv2
-	./pianoteq --export-lv2-presets $ZYNTHIAN_MY_DATA_DIR/presets/lv2
+if [[ "$VIRTUALIZATION" == "none" ]]; then
+	version=$(./pianoteq --version | cut -d' ' -f4)
+	if [[ $version > "7.2.0" ]]; then
+		rm -rf $ZYNTHIAN_MY_DATA_DIR/presets/lv2/"Pianoteq 7 "*"-factory-presets"*.lv2
+		./pianoteq --export-lv2-presets $ZYNTHIAN_MY_DATA_DIR/presets/lv2
+	fi
 fi
+
+#----------------------------------------------------------
